--- a/src/buffer.rs
+++ b/src/buffer.rs
@@ -63,7 +63,7 @@ impl SamplesBuffer {
 
 impl Source for SamplesBuffer {
     #[inline]
-    fn current_span_len(&self) -> Option<usize> {
+    fn parameters_changed(&self) -> bool {
         None
     }