--- a/src/decoder/flac.rs
+++ b/src/decoder/flac.rs
@@ -77,7 +77,7 @@ where
     R: Read + Seek,
 {
     #[inline]
-    fn current_span_len(&self) -> Option<usize> {
+    fn parameters_changed(&self) -> bool {
         None
     }