--- a/src/decoder/flac.rs
+++ b/src/decoder/flac.rs
@@ -78,7 +78,7 @@ where
 {
     #[inline]
     fn parameters_changed(&self) -> bool {
-        None
+        todo!()
     }
 
     #[inline]