--- a/src/decoder/wav.rs
+++ b/src/decoder/wav.rs
@@ -147,7 +147,7 @@ where
     R: Read + Seek,
 {
     #[inline]
-    fn current_span_len(&self) -> Option<usize> {
+    fn parameters_changed(&self) -> bool {
         None
     }