--- a/src/decoder/wav.rs
+++ b/src/decoder/wav.rs
@@ -148,7 +148,8 @@ where
 {
     #[inline]
     fn parameters_changed(&self) -> bool {
-        None
+        todo!()
+        // None
     }
 
     #[inline]