--- a/src/mixer.rs
+++ b/src/mixer.rs
@@ -76,7 +76,7 @@ pub struct MixerSource {
 
 impl Source for MixerSource {
     #[inline]
-    fn current_span_len(&self) -> Option<usize> {
+    fn parameters_changed(&self) -> bool {
         None
     }