--- a/src/mixer.rs
+++ b/src/mixer.rs
@@ -77,7 +77,8 @@ pub struct MixerSource {
 impl Source for MixerSource {
     #[inline]
     fn parameters_changed(&self) -> bool {
-        None
+        todo!()
+        // None
     }
 
     #[inline]