--- a/src/source/channel_volume.rs
+++ b/src/source/channel_volume.rs
@@ -102,8 +102,8 @@ where
     I: Source,
 {
     #[inline]
-    fn current_span_len(&self) -> Option<usize> {
-        self.input.current_span_len()
+    fn parameters_changed(&self) -> bool {
+        self.input.parameters_changed()
     }
 
     #[inline]