--- a/src/source/empty.rs
+++ b/src/source/empty.rs
@@ -35,7 +35,7 @@ impl Iterator for Empty {
 
 impl Source for Empty {
     #[inline]
-    fn current_span_len(&self) -> Option<usize> {
+    fn parameters_changed(&self) -> bool {
         None
     }