--- a/src/source/empty_callback.rs
+++ b/src/source/empty_callback.rs
@@ -32,7 +32,7 @@ impl Iterator for EmptyCallback {
 
 impl Source for EmptyCallback {
     #[inline]
-    fn current_span_len(&self) -> Option<usize> {
+    fn parameters_changed(&self) -> bool {
         None
     }