--- a/src/source/empty_callback.rs
+++ b/src/source/empty_callback.rs
@@ -33,7 +33,7 @@ impl Iterator for EmptyCallback {
 impl Source for EmptyCallback {
     #[inline]
     fn parameters_changed(&self) -> bool {
-        None
+        false
     }
 
     #[inline]