--- a/src/source/linear_ramp.rs
+++ b/src/source/linear_ramp.rs
@@ -108,8 +108,8 @@ where
     I: Source,
 {
     #[inline]
-    fn current_span_len(&self) -> Option<usize> {
-        self.input.current_span_len()
+    fn parameters_changed(&self) -> bool {
+        self.input.parameters_changed()
     }
 
     #[inline]