--- a/src/source/mix.rs
+++ b/src/source/mix.rs
@@ -80,9 +80,9 @@ where
     I2: Source,
 {
     #[inline]
-    fn current_span_len(&self) -> Option<usize> {
-        let f1 = self.input1.current_span_len();
-        let f2 = self.input2.current_span_len();
+    fn parameters_changed(&self) -> bool {
+        let f1 = self.input1.parameters_changed();
+        let f2 = self.input2.parameters_changed();
 
         match (f1, f2) {
             (Some(f1), Some(f2)) => Some(cmp::min(f1, f2)),