--- a/src/source/periodic.rs
+++ b/src/source/periodic.rs
@@ -100,8 +100,8 @@ where
     F: FnMut(&mut I),
 {
     #[inline]
-    fn current_span_len(&self) -> Option<usize> {
-        self.input.current_span_len()
+    fn parameters_changed(&self) -> bool {
+        self.input.parameters_changed()
     }
 
     #[inline]