--- a/src/source/sawtooth.rs
+++ b/src/source/sawtooth.rs
@@ -39,7 +39,7 @@ impl Iterator for SawtoothWave {
 
 impl Source for SawtoothWave {
     #[inline]
-    fn current_span_len(&self) -> Option<usize> {
+    fn parameters_changed(&self) -> bool {
         None
     }