--- a/src/source/sawtooth.rs
+++ b/src/source/sawtooth.rs
@@ -40,7 +40,7 @@ impl Iterator for SawtoothWave {
 impl Source for SawtoothWave {
     #[inline]
     fn parameters_changed(&self) -> bool {
-        None
+        false
     }
 
     #[inline]