--- a/src/source/signal_generator.rs
+++ b/src/source/signal_generator.rs
@@ -137,7 +137,7 @@ impl Iterator for SignalGenerator {
 
 impl Source for SignalGenerator {
     #[inline]
-    fn current_span_len(&self) -> Option<usize> {
+    fn parameters_changed(&self) -> bool {
         None
     }