implements #694

currently draft, due to high number of non trivial changes this PR also aims to dramatically extend the test suite of rodio. Hopefully that will catch most bugs.

@roderickvd I could use some help fixing the decoders, they now just have `todo!()` as `parameters_changed` implementation. Let me know if you want to work on that. The branch is under the rodio repo so you can just pull it down and push your work to it.

status (ill edit this along the way)
- [x] trivial sources implemented (empty, sine etc)
- [x] normal sources implemented (skip etc)
- [ ] extra tests for normal sources
- [ ] complex sources implemented (queue and friends) 
- [ ] extra tests for complex sources
- [ ] decoder sources implemented
- [ ] extra tests for decoder sources (unsure if needed?)