@dvdsk 2025-02-28
first benchmark data is in. Key takeaways:
- resampler is 1.3x slower. (needs to be looked into)
- high_pass is 3.8x faster :partying_face: 
- reverb is 1.5x slower, probably a bad `buffered` implementation. (needs to be investigated)

I'd not expect those changes to be honest. Lets hope we can speed things up otherwise we have some hard choices to make. Lets not dwell on it, I have spend literally zero time on optimization and I have not written this with performance in mind (yet).

### Current master
```
Timer precision: 20 ns
conversions     fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ from_sample                │               │               │               │         │
   ├─ f32       564.8 µs      │ 1.377 ms      │ 737 µs        │ 761.8 µs      │ 100     │ 100
   ├─ i16       254.5 µs      │ 360.5 µs      │ 255.9 µs      │ 257.5 µs      │ 100     │ 100
   ╰─ u16       380 µs        │ 479.2 µs      │ 382.2 µs      │ 383.9 µs      │ 100     │ 100

     Running benches/effects.rs (target/release/deps/effects-b27671fcc5098510)
Timer precision: 20 ns
effects         fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ agc_enabled  3.681 ms      │ 4.98 ms       │ 3.704 ms      │ 3.74 ms       │ 100     │ 100
├─ amplify      382.1 µs      │ 419.6 µs      │ 382.2 µs      │ 384.5 µs      │ 100     │ 100
├─ fade_out     1.142 ms      │ 1.161 ms      │ 1.148 ms      │ 1.148 ms      │ 100     │ 100
├─ high_pass    6.656 ms      │ 7.304 ms      │ 6.675 ms      │ 6.685 ms      │ 100     │ 100
╰─ reverb       8.281 ms      │ 8.422 ms      │ 8.322 ms      │ 8.326 ms      │ 100     │ 100

     Running benches/pipeline.rs (target/release/deps/pipeline-bb11d51b2daeec9e)
Timer precision: 30 ns
pipeline  fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ long   26.06 ms      │ 27.09 ms      │ 26.15 ms      │ 26.19 ms      │ 100     │ 100
╰─ short  3.788 ms      │ 3.856 ms      │ 3.839 ms      │ 3.835 ms      │ 100     │ 100

     Running benches/resampler.rs (target/release/deps/resampler-9a89b0f001b3c0cb)
Timer precision: 20 ns
resampler         fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ no_resampling  2.762 ms      │ 2.991 ms      │ 2.799 ms      │ 2.803 ms      │ 100     │ 100
╰─ resample_to                  │               │               │               │         │
   ├─ 8000        2.739 ms      │ 3.65 ms       │ 2.761 ms      │ 2.769 ms      │ 100     │ 100
   ├─ 11025       3.216 ms      │ 3.324 ms      │ 3.234 ms      │ 3.235 ms      │ 100     │ 100
   ├─ 16000       3.801 ms      │ 3.95 ms       │ 3.829 ms      │ 3.83 ms       │ 100     │ 100
   ├─ 22050       4.628 ms      │ 4.816 ms      │ 4.795 ms      │ 4.774 ms      │ 100     │ 100
   ├─ 44100       2.769 ms      │ 2.901 ms      │ 2.792 ms      │ 2.793 ms      │ 100     │ 100
   ├─ 48000       7.135 ms      │ 7.357 ms      │ 7.23 ms       │ 7.224 ms      │ 100     │ 100
   ├─ 88200       10.82 ms      │ 11.01 ms      │ 10.9 ms       │ 10.9 ms       │ 100     │ 100
   ├─ 96000       12.2 ms       │ 12.62 ms      │ 12.35 ms      │ 12.36 ms      │ 100     │ 100
   ├─ 176400      20.32 ms      │ 20.92 ms      │ 20.52 ms      │ 20.55 ms      │ 100     │ 100
   ├─ 192000      22.71 ms      │ 23.12 ms      │ 22.88 ms      │ 22.89 ms      │ 100     │ 100
   ├─ 352800      38.86 ms      │ 39.6 ms       │ 38.92 ms      │ 38.94 ms      │ 100     │ 100
   ╰─ 384000      45 ms         │ 47.65 ms      │ 45.07 ms      │ 45.15 ms      │ 100     │ 100
```

### This PR
```
Timer precision: 20 ns
conversions     fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ from_sample                │               │               │               │         │
   ├─ f32       753.3 µs      │ 969 µs        │ 919.8 µs      │ 920.8 µs      │ 100     │ 100
   ├─ i16       255 µs        │ 373.2 µs      │ 255.3 µs      │ 258.4 µs      │ 100     │ 100
   ╰─ u16       380 µs        │ 390.1 µs      │ 380.2 µs      │ 381.7 µs      │ 100     │ 100

     Running benches/effects.rs (target/release/deps/effects-4d762cfcb96f6e43)
Timer precision: 20 ns
effects         fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ agc_enabled  3.695 ms      │ 4.135 ms      │ 3.699 ms      │ 3.709 ms      │ 100     │ 100
├─ amplify      380 µs        │ 398.3 µs      │ 380.4 µs      │ 382.5 µs      │ 100     │ 100
├─ fade_out     1.141 ms      │ 1.224 ms      │ 1.142 ms      │ 1.146 ms      │ 100     │ 100
├─ high_pass    1.711 ms      │ 1.733 ms      │ 1.714 ms      │ 1.715 ms      │ 100     │ 100
╰─ reverb       12.65 ms      │ 13.03 ms      │ 12.68 ms      │ 12.69 ms      │ 100     │ 100

     Running benches/pipeline.rs (target/release/deps/pipeline-113220290e55aa6e)
Timer precision: 20 ns
pipeline  fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ long   30.83 ms      │ 34.99 ms      │ 30.87 ms      │ 30.92 ms      │ 100     │ 100
╰─ short  1.711 ms      │ 1.73 ms       │ 1.714 ms      │ 1.714 ms      │ 100     │ 100

     Running benches/resampler.rs (target/release/deps/resampler-6724b60d4573a4a8)
Timer precision: 20 ns
resampler         fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ no_resampling  3.819 ms      │ 4.011 ms      │ 3.83 ms       │ 3.833 ms      │ 100     │ 100
╰─ resample_to                  │               │               │               │         │
   ├─ 8000        3.196 ms      │ 3.71 ms       │ 3.584 ms      │ 3.552 ms      │ 100     │ 100
   ├─ 11025       3.244 ms      │ 6.276 ms      │ 3.63 ms       │ 3.651 ms      │ 100     │ 100
   ├─ 16000       4.125 ms      │ 4.529 ms      │ 4.479 ms      │ 4.474 ms      │ 100     │ 100
   ├─ 22050       4.709 ms      │ 4.764 ms      │ 4.721 ms      │ 4.726 ms      │ 100     │ 100
   ├─ 44100       3.816 ms      │ 3.936 ms      │ 3.832 ms      │ 3.869 ms      │ 100     │ 100
   ├─ 48000       8.168 ms      │ 8.327 ms      │ 8.277 ms      │ 8.275 ms      │ 100     │ 100
   ├─ 88200       11.43 ms      │ 12.03 ms      │ 11.49 ms      │ 11.56 ms      │ 100     │ 100
   ├─ 96000       13.98 ms      │ 14.23 ms      │ 14.16 ms      │ 14.17 ms      │ 100     │ 100
   ├─ 176400      21.14 ms      │ 22.38 ms      │ 21.29 ms      │ 21.35 ms      │ 100     │ 100
   ├─ 192000      25.14 ms      │ 25.84 ms      │ 25.43 ms      │ 25.47 ms      │ 100     │ 100
   ├─ 352800      41.62 ms      │ 43.17 ms      │ 42.3 ms       │ 42.25 ms      │ 100     │ 100
   ╰─ 384000      46.84 ms      │ 67.99 ms      │ 48.41 ms      │ 48.47 ms      │ 100     │ 100
```